Added try catch error handling code in app/Models folder files except the world folder files.

+++ app/Models/getmagazinedata.php
@@ -1,38 +1,55 @@
 <?php
-   require '../app/Config/srpss_database_connection.php';
-   if(isset($error))
+   try
    {
-      echo $error;
-      exit();
-   }
-   // store query in a variable
-   $sql = "SELECT * FROM magazines";
-
-   // perform query and store result in a variable
-   $result = $conn->query($sql);
-
-   if($result->num_rows > 0)
-   {
-      echo "<table>
-      <tr>
-      <th>S.N.</th>
-      <th>Name</th>
-      <th>Publisher</th>
-      </tr>";
-
-      while($row = $result->fetch_assoc()) {
-         echo "<tr>";
-         echo "<td>" . $row['id'] . "</td>";
-         echo "<td>" . $row['magazinename'] . "</td>";
-         echo "<td>" . $row['publisher'] . "</td>";
-         echo "</tr>";
+      if(!file_exists('../app/Config/srpss_database_connection.php'))
+      {
+         throw new Exception("srpss_database_connection.php is missing.");
       }
-      echo "</table>";
+      else
+      {
+         require '../app/Config/srpss_database_connection.php';
+         if(isset($error))
+         {
+            echo "<script>alert($error);</script>";
+            //exit();
+         }
+         else
+         {
+            // store query in a variable
+            $sql = "SELECT * FROM magazines";
+
+            // perform query and store result in a variable
+            $result = $conn->query($sql);
+
+            if($result->num_rows > 0)
+            {
+               echo "<table>
+               <tr>
+               <th>S.N.</th>
+               <th>Name</th>
+               <th>Publisher</th>
+               </tr>";
+
+               while($row = $result->fetch_assoc()) {
+                  echo "<tr>";
+                  echo "<td>" . $row['id'] . "</td>";
+                  echo "<td>" . $row['magazinename'] . "</td>";
+                  echo "<td>" . $row['publisher'] . "</td>";
+                  echo "</tr>";
+               }
+               echo "</table>";
+            }
+            else
+            {
+               echo "<p style='color:red;'>No Magazine data found.</p>";
+            }
+            // close the connection
+            $conn->close();
+         }
+      }
    }
-   else
+   catch(Exception $e)
    {
-      echo "<p style='color:red;'>No Magazine data found.</p>";
+      echo "<script>alert('{$e->getMessage()}');</script>";
    }
-   // close the connection
-   $conn->close();
 ?>
